+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/job/commit/CompactionJobCommitRequest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.job.completion;
+package sleeper.compaction.job.commit;
 
 import sleeper.compaction.job.CompactionJob;
 import sleeper.core.record.process.RecordsProcessed;
