+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/job/commit/CompactionJobCommitRequestSerDeTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.job.completion;
+package sleeper.compaction.job.commit;
 
 import org.approvaltests.Approvals;
 import org.junit.jupiter.api.Test;
